boatsphotosController & data & business services added

+++ Boat.Data/DataModel/BoatModule/Interface/IBoatPhotosRepository.cs
@@ -1,10 +1,9 @@
 ﻿using Boat.Data.DataModel.BoatModule.Entity;
 using Boat.Data.Interface;
-using System.Collections.Generic;
 
 namespace Boat.Data.DataModel.BoatModule.Interface
 {
-    public interface IBoatPhotosRepository : IRepository<BoatPhotos, int>, IRepository
+    public interface IBoatPhotosRepository : IRepository<BoatPhotos, long>, IRepository
     {
     }
 }
